Fixed NEON compilation issue
